refactor: Moved searchAndParse and searchDirectory method to new CitationFinder class
